+++ docs/docs/30-administration/75-addons/20-creating-addons.md
@@ -23,10 +23,6 @@ Directly import Woodpecker's Go package (`go.woodpecker-ci.org/woodpecker/woodpe
 | -------------------- | ------------------------------------------------------------------------------------- |
 | `Forge`              | `"go.woodpecker-ci.org/woodpecker/woodpecker/v2/server/forge".Forge`                  |
 | `Backend`            | `"go.woodpecker-ci.org/woodpecker/woodpecker/v2/pipeline/backend/types".Backend`      |
-| `ConfigService`      | `"go.woodpecker-ci.org/woodpecker/v2/server/services/config".Extension`               |
-| `SecretService`      | `"go.woodpecker-ci.org/woodpecker/v2/server/services/secret".SecretService`           |
-| `EnvironmentService` | `"go.woodpecker-ci.org/woodpecker/v2/server/services/environment".EnvironmentService` |
-| `RegistryService`    | `"go.woodpecker-ci.org/woodpecker/v2/server/services/registry".RegistryService`       |
 
 ### Using configurations
