+++ server/extensions/config/http.go
@@ -19,10 +19,10 @@ import (
 	"crypto"
 	"fmt"
 
+	"go.woodpecker-ci.org/woodpecker/v2/server/extensions/utils"
 	"go.woodpecker-ci.org/woodpecker/v2/server/forge"
 	"go.woodpecker-ci.org/woodpecker/v2/server/forge/types"
 	"go.woodpecker-ci.org/woodpecker/v2/server/model"
-	"go.woodpecker-ci.org/woodpecker/v2/server/services/utils"
 )
 
 type http struct {
