+++ server/extensions/manager.go
@@ -12,18 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package services
+package extensions
 
 import (
 	"crypto"
 
 	"github.com/urfave/cli/v2"
 
+	"go.woodpecker-ci.org/woodpecker/v2/server/extensions/config"
+	"go.woodpecker-ci.org/woodpecker/v2/server/extensions/environment"
+	"go.woodpecker-ci.org/woodpecker/v2/server/extensions/registry"
+	"go.woodpecker-ci.org/woodpecker/v2/server/extensions/secret"
 	"go.woodpecker-ci.org/woodpecker/v2/server/model"
-	"go.woodpecker-ci.org/woodpecker/v2/server/services/config"
-	"go.woodpecker-ci.org/woodpecker/v2/server/services/environment"
-	"go.woodpecker-ci.org/woodpecker/v2/server/services/registry"
-	"go.woodpecker-ci.org/woodpecker/v2/server/services/secret"
 	"go.woodpecker-ci.org/woodpecker/v2/server/store"
 )
