+++ server/extensions/registry/combined.go
