+++ server/extensions/registry/service.go
